\name{for.each}
\alias{for.each}
\title{Apply \code{f} to the successive elements of \code{...}.}
\usage{
  for.each(f, ...)
}
\arguments{
  \item{f}{the function to apply, whose arity should match
  the cardinality of \code{...}}

  \item{...}{lists upon which to apply \code{f}
  successively}
}
\value{
  NULL
}
\description{
  Apply \code{f} to the successive elements of \code{...}.
}

