\name{zip}
\alias{zip}
\title{Zip \emph{n} lists together into tuplets of
length \emph{n}.}
\usage{
  zip(zipper, ...)
}
\arguments{
  \item{zipper}{the zipping function}

  \item{\dots}{the lists to be zipped}
}
\value{
  A list of tuplets
}
\description{
  Zip \emph{n} lists together into tuplets of length
  \emph{n}.
}

