\name{zip.with.names}
\alias{zip.with.names}
\title{Do a less efficient zip whilst preserving names.}
\usage{
  zip.with.names(...)
}
\arguments{
  \item{...}{lists to be zipped whilst preserving names}
}
\value{
  A list of tuples (one per position) with names preserved from the
  input lists.
}
\description{
  Do a less efficient zip whilst preserving names.
}

