\name{for.each}
\alias{for.each}
\title{Apply f to the successive elements of .}
\usage{for.each(f, ...)}
\description{Apply \code{f} to the successive elements of \code{...}.}
\value{NULL}
\arguments{\item{f}{the function to apply, whose arity should match the
cardinality of \code{...}}
\item{...}{lists upon which to apply \code{f} successively}}
