\name{pair.fold.right}
\alias{pair.fold.right}
\title{pair-fold-right from SRFI-1.}
\usage{pair.fold.right(f, nil, ...)}
\description{pair-fold-right from SRFI-1.}
\section{TODO}{one-list fast-path}
\arguments{\item{f}{function to apply over the list-tails}
\item{nil}{the default value}
\item{...}{the lists whose tails fold over}}
